Extract ModifiedMethodsHelper
