+++ src/services/modifiedLinesCollector/ModifiedLinesCollector.groovy
@@ -17,10 +17,7 @@ class ModifiedLinesCollector implements DataCollector {
     private File experimentalDataFile;
     private File experimentalDataFileWithLinks;
 
-    private TextualDiffParser textualDiffParser = new TextualDiffParser();
-    private DiffJParser modifiedMethodsParser = new DiffJParser();
-    private MethodModifiedLinesMatcher modifiedMethodsMatcher = new MethodModifiedLinesMatcher();
-
+    private ModifiedMethodsHelper modifiedMethodsHelper = new ModifiedMethodsHelper();
     private RevisionsFilesCollector revisionsCollector = new RevisionsFilesCollector();
     
     private void createOutputFiles(String outputPath) {
@@ -46,7 +43,7 @@ class ModifiedLinesCollector implements DataCollector {
 
         for (String filePath : mutuallyModifiedFiles) {
             // get merge revision modified methods
-            Set<ModifiedMethod> allModifiedMethods = getModifiedMethods(project, filePath, mergeCommit.getAncestorSHA(), mergeCommit.getSHA())
+            Set<ModifiedMethod> allModifiedMethods = modifiedMethodsHelper.getModifiedMethods(project, filePath, mergeCommit.getAncestorSHA(), mergeCommit.getSHA())
             // get methods modified by both left and right revisions
             Map<String, Tuple2<ModifiedMethod, ModifiedMethod>> mutuallyModifiedMethods = getMutuallyModifiedMethods(project, mergeCommit, filePath);
 
@@ -90,6 +87,7 @@ class ModifiedLinesCollector implements DataCollector {
                             }
                         }
 
+                        // prints results to a csv file
                         printResults(project, mergeCommit, className, method.getSignature(), leftAddedLines, leftDeletedLines, rightAddedLines,rightDeletedLines); 
                     }
 
@@ -106,7 +104,6 @@ class ModifiedLinesCollector implements DataCollector {
                       HashSet<Integer> leftAddedLines, HashSet<Integer> leftDeletedLines, HashSet<Integer> rightAddedLines,
                       HashSet<Integer> rightDeletedLines) {
 
-
         experimentalDataFile << "${project.getName()};${mergeCommit.getSHA()};${className};${modifiedDeclarationSignature};${leftAddedLines};${leftDeletedLines};${rightAddedLines};${rightDeletedLines}\n"
 
         // Add links.
@@ -115,10 +112,16 @@ class ModifiedLinesCollector implements DataCollector {
 
     }
 
+    private Set<String> getMutuallyModifiedFiles(Project project, MergeCommit mergeCommit) {
+        Set<String> leftModifiedFiles = FileManager.getModifiedFiles(project, mergeCommit.getLeftSHA(), mergeCommit.getAncestorSHA())
+        Set<String> rightModifiedFiles = FileManager.getModifiedFiles(project, mergeCommit.getRightSHA(), mergeCommit.getAncestorSHA())
+
+        return leftModifiedFiles.intersect(rightModifiedFiles)
+    }
 
     private Map<String, Tuple2<ModifiedMethod, ModifiedMethod>> getMutuallyModifiedMethods(Project project, MergeCommit mergeCommit, String filePath) {
-        Set<ModifiedMethod> leftModifiedMethods = getModifiedMethods(project, filePath, mergeCommit.getAncestorSHA(), mergeCommit.getLeftSHA())
-        Set<ModifiedMethod> rightModifiedMethods = getModifiedMethods(project, filePath, mergeCommit.getAncestorSHA(), mergeCommit.getRightSHA())
+        Set<ModifiedMethod> leftModifiedMethods = modifiedMethodsHelper.getModifiedMethods(project, filePath, mergeCommit.getAncestorSHA(), mergeCommit.getLeftSHA())
+        Set<ModifiedMethod> rightModifiedMethods = modifiedMethodsHelper.getModifiedMethods(project, filePath, mergeCommit.getAncestorSHA(), mergeCommit.getRightSHA())
         return intersectAndBuildMap(leftModifiedMethods, rightModifiedMethods)
     }
 
@@ -135,39 +138,4 @@ class ModifiedLinesCollector implements DataCollector {
 
         return intersection
     }    
-    private Set<String> getMutuallyModifiedFiles(Project project, MergeCommit mergeCommit) {
-        Set<String> leftModifiedFiles = FileManager.getModifiedFiles(project, mergeCommit.getLeftSHA(), mergeCommit.getAncestorSHA())
-        Set<String> rightModifiedFiles = FileManager.getModifiedFiles(project, mergeCommit.getRightSHA(), mergeCommit.getAncestorSHA())
-
-        return leftModifiedFiles.intersect(rightModifiedFiles)
-    }
-
-    private Set<ModifiedMethod> getModifiedMethods(Project project, String filePath, String ancestorSHA, String targetSHA) {
-        File ancestorFile = FileManager.getFileInCommit(project, filePath, ancestorSHA)
-        File targetFile = FileManager.getFileInCommit(project, filePath, targetSHA)
-
-        List<String> diffJOutput = runDiffJ(ancestorFile, targetFile);
-        List<String> textualDiffOutput = runTextualDiff(ancestorFile, targetFile);
-
-        Map<String, int[]> parsedDiffJResult = modifiedMethodsParser.parse(diffJOutput);
-        List<ModifiedLine> parsedTextualDiffResult = textualDiffParser.parse(textualDiffOutput);
-
-        return modifiedMethodsMatcher.matchModifiedMethodsAndLines(parsedDiffJResult, parsedTextualDiffResult);
-    }
-
-    private List<String> runDiffJ(File ancestorFile, File targetFile) {
-        Process diffJ = ProcessRunner.runProcess('dependencies', 'java', '-jar', 'diffj.jar', "--brief", ancestorFile.getAbsolutePath(), targetFile.getAbsolutePath())
-        BufferedReader reader = new BufferedReader(new InputStreamReader(diffJ.getInputStream()))
-        def output = reader.readLines()
-        reader.close()
-        return output
-    }
-
-    private List<String> runTextualDiff (File ancestorFile, File targetFile) {
-        Process textDiff = ProcessRunner.runProcess(".", "diff" ,ancestorFile.getAbsolutePath(), targetFile.getAbsolutePath())
-        BufferedReader reader = new BufferedReader(new InputStreamReader(textDiff.getInputStream()))
-        def output = reader.readLines()
-        reader.close()
-        return output
-    }
 }
